refactor(styles): wrappers module created
