+++ app/components/wrapper/main.tsx
@@ -1,5 +1,5 @@
 import type { PropsWithChildren } from 'react';
 
-export const Main = ({ children }: PropsWithChildren) => (
+export const MainWrapper = ({ children }: PropsWithChildren) => (
     <main className="min-h-screen px-8">{children}</main>
 );
